mseis/requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.1.0
langchain-openai==0.0.5
langchain-experimental==0.0.47
pinecone-client==3.0.0
neo4j==5.15.0
openai==1.10.0
anthropic==0.12.0

# Vector and ML
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
clip-interrogator==0.6.0
faiss-cpu==1.7.4

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
arxiv==2.0.0
pillow==10.1.0
opencv-python==4.8.1

# API and Web
requests==2.31.0
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0

# Streaming and Real-time
asyncio==3.4.3
websockets==12.0
sse-starlette==1.8.2

# UI and Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
gradio==4.14.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development Tools
python-dotenv==1.0.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Caching and Performance
redis==5.0.1
diskcache==5.6.3
cachetools==5.3.2

# Rate Limiting
ratelimit==2.2.1
slowapi==0.1.9

# Additional dependencies
rank_bm25==0.2.2
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1
PyYAML==6.0.1
feedparser==6.0.10 